move session state keys to dedicated module
