+++ tests/gui/test_02_import_data.py
@@ -2,7 +2,7 @@ from unittest.mock import MagicMock, patch
 
 from streamlit.testing.v1 import AppTest
 
-from alphastats.gui.utils.ui_helper import StateKeys
+from alphastats.gui.utils.state_keys import StateKeys
 
 from .conftest import APP_FOLDER, data_buf, metadata_buf
